Create 5 modules from the main js file
