+++ index.html
@@ -4,9 +4,9 @@
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="style.css">
+    <link rel="stylesheet" href="./index.css">
     <link href='https://fonts.googleapis.com/css?family=Poppins' rel='stylesheet'>
-    <script src="./book.js"></script>
+    <script type="module" src="./index.js"></script>
 
     <title>Iread Book Store </title>
 </head>
@@ -18,9 +18,9 @@
              <img  alt="logo" src="book-logo.png" alt="awosome book logo"/> 
             </div>
             <ul>
-                <li><a href="#" onclick="getPageContent('list')" id="btnList">List</a></li>
-                <li><a href="#" onclick="getPageContent('addNew')" id="btnContact'">Add new</a></li>
-                <li><a href="#" onclick="getPageContent('contact')" id="btnContact">Contact</a></li>
+                <li><a href="#"  id="btnList">List</a></li>
+                <li><a href="#" id="btnAddbook">Add new</a></li>
+                <li><a href="#"  id="btnContact">Contact</a></li>
             </ul>
         </nav>
     </header>
