+++ modules/contact.js
@@ -1,0 +1,17 @@
+/* eslint-disable import/prefer-default-export */
+export const contact = () => {
+  const content1 = `<div class="contact-continar">
+      <h2>Contact information</h2>
+      <div class="contact-content">
+      <p>Do you have any questions<br>
+      You can reach out to us!
+      </p>
+      <ul>
+      <li> Our e-mail<a>misikirteka@gmail.com</a></li>
+      <li>Our phone number:+251930955432 </li>
+      <li> Our address : Microverse Students</li>
+      </ul>
+      </div>
+      </div>`;
+  return content1;
+};
